v9.8: SIGTRAP fix in encodeViaQstepSearch + corpus calibration data
Real production bug fix: encodeViaQstepSearch's post-loop refinement
crashed with SIGTRAP when called from encodeWithQstepStats (which
passes maxOvershootRatio=.infinity). The trap is in:

    while bestEncoded.count > Int(targetBytes * maxOvershootRatio),
          refinementIters < maxRefinementIters {

Int(Double.infinity) is undefined behaviour in Swift and traps.
Triggered only when the main qstep search exits without converging
within maxIterations (high-variance content where the bracket
doesn't reach the true converged qstep in 8 iter). Smaller fixtures
(mr_002) avoided it because they converged in the main loop. Larger
fixtures (ct_001 / dx_002 / px_001 / mg_001 / mg_002) crashed
reliably.

Fix: compare in Double space so .infinity short-circuits cleanly.
Double(count) > .infinity is always false, disabling the refinement
loop in the infinite-cap case.

Test impact: J2KCrossScaleRDQualityProbe's three test methods
(testCrossScaleQstepVsPCRD, testCrossScaleRDPSNRSweep,
testDX002LosslessRoundtripExact) all previously SIGTRAPped and now
pass cleanly. The v9.7 doc filed this as "pre-existing failure" —
it was the qstep-search refinement bug exposed when the search
maxed out iterations on high-variance content.

Also unblocked the V98QstepCalibrationTests suite for the full
corpus. With the SIGTRAP fixed, gathered tight-tolerance qstep
search data for all 7 corpus fixtures at 2 bpp. Key finding: only
mr_002 (smallest, mid-texture) converges within ±5% in 8 iter.
Larger fixtures exit at the bracket cap, revealing that the
search's bracket widening is too slow for high-variance content —
a SECOND issue distinct from the v9.7 pre-seed problem.

V9_8_IMAGE_STATS_RESEARCH.md updated with:
- SIGTRAP fix methodology + test impact
- Full corpus stats (texture, mean_diff, pixels)
- Tight-tolerance qstep convergence results
- Analysis of why most fixtures don't converge in 8 iter (bracket
  widening is sub-linear vs. the required qstep range)
- Multi-feature predictor design implications for future v9.9 work

Bit-exact gates pass (HTCrossCodec MD5 parity, V94NEONHotPath,
V91Phase2c, HTBlockEncoderConformant, HTConformantConstantBitrateViaQstep,
HTConformantFixedQstepRD, J2KCrossScaleRDQualityProbe, V91Phase2-
ConcurrentContentionProbe).

+++ Sources/J2KCodec/J2KCodec.swift
@@ -853,7 +853,14 @@ public struct J2KEncoder: Sendable {
         // conformant cleanup-only block format.
         var refinementIters = 0
         let maxRefinementIters = 3
-        while bestEncoded.count > Int(targetBytes * maxOvershootRatio),
+        // v9.8 — pre-existing bug fix: when `maxOvershootRatio` is
+        // `.infinity` (the path taken by `encodeWithQstepStats`),
+        // `Int(targetBytes * .infinity)` traps with SIGTRAP. Comparing
+        // in Double space lets `infinity` short-circuit cleanly:
+        // `Double(count) > .infinity` is always false, so the
+        // refinement loop is correctly disabled in the infinite-cap
+        // case.
+        while Double(bestEncoded.count) > targetBytes * maxOvershootRatio,
               refinementIters < maxRefinementIters {
             refinementIters += 1
             let currentRatio = Double(bestEncoded.count) / targetBytes
